Refactor
Remove old unused files. Reorganized files in different folders
