+++ jaqpotpy/descriptors/graph/__init__.py
@@ -1,4 +1,1 @@
-from jaqpotpy.descriptors.graph.graph_data import *
-from jaqpotpy.descriptors.graph.graph_features import *
-from jaqpotpy.descriptors.graph.mol_graphs import *
 from jaqpotpy.descriptors.graph.graph_featurizer import *
