+++ jaqpotpy/models/torch_geometric_models/__init__.py
@@ -1,12 +1,4 @@
-# from .graph_attention_network import GraphAttentionNetwork
-from .graph_convolutional_network import GraphConvolutionNetwork
-from .graph_sage_network import GraphSageNetwork
-
-# from .graph_transformer_network import GraphTransformerNetwork
-
-# from .fully_connected_network import FullyConnectedNetwork
-
-# from .graph_attention_network import GraphAttentionNetworkWithExternal
-# from .graph_convolutional_network import GraphConvolutionalNetworkWithExternal
-# from .graph_sage_network import GraphSAGENetworkWithExternal
-# from .graph_transformer_network import GraphTransformerNetworkWithExternal
+from jaqpotpy.models.torch_geometric_models.graph_neural_network import (
+    GraphConvolutionNetwork,
+    GraphSageNetwork,
+)
