--- jaqpotpy/models/trainers/base/__init__.py
@@ -1,5 +1,0 @@
-"""
-Author: Ioannis Pitoskas (jpitoskas@gmail.com)
-"""
-
-from .torch_model_trainer import TorchModelTrainer
