+++ jaqpotpy/models/trainers/base_trainer.py
