+++ jaqpotpy/models/trainers/binary_trainer.py
@@ -1,15 +1,30 @@
-"""Author: Ioannis Pitoskas (jpitoskas@gmail.com)"""
-
-from ..base import TorchModelTrainer
-from abc import abstractmethod
+from jaqpotpy.api.openapi.jaqpot_api_client.models.feature import Feature
+from jaqpotpy.api.openapi.jaqpot_api_client.models.feature_type import FeatureType
 from tqdm import tqdm
 import torch
+import io
+import base64
+from typing import Optional
+import inspect
+from jaqpotpy.models.trainers import TorchModelTrainer
+from sklearn import metrics
 import torch.nn.functional as F
-import sklearn.metrics as metrics
 
 
-class MulticlassModelTrainer(TorchModelTrainer):
-    """Abstract trainer class for Multiclass Classification models using PyTorch."""
+class BinaryGraphModelTrainer(TorchModelTrainer):
+    """
+    Trainer class for Binary Classification using Graph Neural Networks for SMILES and external features.
+
+    Attributes:
+        decision_threshold (float): Decision threshold for binary classification.
+    """
+
+    MODEL_TYPE = "binary-graph-model"
+    """'binary-graph-model'"""
+
+    @classmethod
+    def get_model_type(cls):
+        return cls.MODEL_TYPE
 
     def __init__(
         self,
@@ -23,10 +38,10 @@ class MulticlassModelTrainer(TorchModelTrainer):
         log_enabled=True,
         log_filepath=None,
     ):
-        """The MulticlassModelTrainer constructor.
+        """
+        The BinaryGraphModelTrainer constructor.
 
         Args:
-        ----
             model (torch.nn.Module): The torch model to be trained.
             n_epochs (int): Number of training epochs.
             optimizer (torch.optim.Optimizer): The optimizer used for training the model.
@@ -36,7 +51,23 @@ class MulticlassModelTrainer(TorchModelTrainer):
             use_tqdm (bool, optional): Whether to use tqdm for progress bars. Default is True.
             log_enabled (bool, optional): Whether logging is enabled. Default is True.
             log_filepath (str or None, optional): Path to the log file. If None, logging is not saved to a file. Default is None.
+            decision_threshold (float, optional): Decision threshold for binary classification. Default is 0.5.
 
+        Example:
+        ```
+        >>> import torch
+        >>> from jaqpotpy.jaqpotpy_torch.models import GraphAttentionNetwork
+        >>> from jaqpotpy.jaqpotpy_torch.trainers import BinaryGraphModelTrainer
+        >>>
+        >>> model = GraphAttentionNetwork(input_dim=10,
+        ...                               hidden_dims=[32, 32]
+        ...                               edge_dim=5,
+        ...                               output_dim=1)
+        >>> optimizer = torch.optim.Adam(model.parameters(), lr=0.0005)
+        >>> loss_fn = torch.nn.BCEWithLogitsLoss()
+        >>>
+        >>> trainer = BinaryGraphModelTrainer(model, n_epochs=50, optimizer=optimizer, loss_fn=loss_fn)
+        ```
         """
         super().__init__(
             model=model,
@@ -50,34 +81,39 @@ class MulticlassModelTrainer(TorchModelTrainer):
             log_filepath=log_filepath,
         )
 
-    @abstractmethod
     def get_model_kwargs(self, data):
-        """This abstract method should be implemented by subclasses to provide model-specific keyword arguments based on the data.
+        """
+        Fetch the model's keyword arguments.
 
         Args:
-        ----
-            data: Whatever data the respective dataloader fetches.
+            data (torch_geometric.data.Data): Data object returned as returned by the Dataloader
 
         Returns:
-        -------
-            dict: The kwargs that the forward method of the respective model expects as input.
-
+            dict: The required model kwargs. Set of keywords: {'*x*', '*edge_index*', '*batch*', '*edge_attr*'}. Note that '*edge_attr*' is only present if the model supports edge features.
         """
-        pass
+
+        kwargs = {}
+
+        kwargs["x"] = data.x
+        kwargs["edge_index"] = data.edge_index
+        kwargs["batch"] = data.batch
+
+        if "edge_attr" in inspect.signature(self.model.forward).parameters:
+            kwargs["edge_attr"] = data.edge_attr
+
+        return kwargs
 
     def train(self, train_loader, val_loader=None):
-        """Train the model.
+        """
+        Train the model.
 
         Args:
-        ----
             train_loader (Union[torch.utils.data.DataLoader, torch_geometric.loader.DataLoader]): DataLoader for the training dataset.
             val_loader (Union[torch.utils.data.DataLoader, torch_geometric.loader.DataLoader], optional): DataLoader for the validation dataset.
-
         Returns:
-        -------
             None
-
         """
+
         for i in range(self.n_epochs):
             self.current_epoch += 1
 
@@ -99,16 +135,13 @@ class MulticlassModelTrainer(TorchModelTrainer):
             self.scheduler.step()
 
     def _train_one_epoch(self, train_loader):
-        """This helper method handles the training loop for a single epoch, updating the model parameters and computing the running loss.
+        """
+        This helper method handles the training loop for a single epoch, updating the model parameters and computing the running loss.
 
         Args:
-        ----
             train_loader (Union[torch.utils.data.DataLoader, torch_geometric.loader.DataLoader]): DataLoader for the training dataset.
-
         Returns:
-        -------
             float: The average loss of the current epoch.
-
         """
         running_loss = 0
         total_samples = 0
@@ -132,8 +165,8 @@ class MulticlassModelTrainer(TorchModelTrainer):
 
             self.optimizer.zero_grad()
 
-            outputs = self.model(**model_kwargs)
-            loss = self.loss_fn(outputs.float(), y.long())
+            outputs = self.model(**model_kwargs).squeeze(-1)
+            loss = self.loss_fn(outputs.float(), y.float())
 
             running_loss += loss.item() * y.size(0)
             total_samples += y.size(0)
@@ -153,20 +186,18 @@ class MulticlassModelTrainer(TorchModelTrainer):
         return avg_loss
 
     def evaluate(self, val_loader):
-        """Evaluate the model's performance on the validation set.
+        """
+        Evaluate the model's performance on the validation set.
 
         Args:
-        ----
             val_loader (Union[torch.utils.data.DataLoader, torch_geometric.loader.DataLoader]): DataLoader for the validation dataset.
-
         Returns:
-        -------
             float: Average loss over the validation dataset.
             dict: Dictionary containing evaluation metrics. The keys represent the metric names and the values are floats.
-            numpy.ndarray: Confusion matrix C as a numpy.ndarray of shape (num_classes, num_classes), with Cij being equal to
-                           the number of observations known to be in group i and predicted to be in group j.
-
+            numpy.ndarray: Confusion matrix as a numpy array of shape (2, 2) representing true negative (TN), false positive (FP),
+                           false negative (FN), and true positive (TP) counts respectively. The elements are arranged as [[TN, FP], [FN, TP]].
         """
+
         running_loss = 0
         total_samples = 0
 
@@ -186,16 +217,16 @@ class MulticlassModelTrainer(TorchModelTrainer):
 
                 model_kwargs = self.get_model_kwargs(data)
 
-                outputs = self.model(**model_kwargs)
+                outputs = self.model(**model_kwargs).squeeze(-1)
 
-                probs = F.softmax(outputs, dim=1)
-                _, preds = torch.max(probs, 1)
+                probs = F.sigmoid(outputs)
+                preds = (probs > 0.5).int()
 
                 all_probs.extend(probs.tolist())
                 all_preds.extend(preds.tolist())
                 all_labels.extend(y.tolist())
 
-                loss = self.loss_fn(outputs.float(), y.long())
+                loss = self.loss_fn(outputs.float(), y.float())
 
                 running_loss += loss.item() * y.size(0)
                 total_samples += y.size(0)
@@ -203,28 +234,20 @@ class MulticlassModelTrainer(TorchModelTrainer):
             avg_loss = running_loss / len(val_loader.dataset)
 
         metrics_dict = self._compute_metrics(all_labels, all_preds)
-        # metrics_dict['roc_auc'] = metrics.roc_auc_score(all_labels, all_probs)
+        metrics_dict["roc_auc"] = metrics.roc_auc_score(all_labels, all_probs)
         metrics_dict["loss"] = avg_loss
-        conf_mat = metrics.confusion_matrix(
-            all_labels, all_preds, labels=torch.arange(self.model.output_dim)
-        )
-
-        #     tn, fp, fn, tp = metrics.confusion_matrix(all_labels, all_preds).ravel()
-        #     conf_mat = {'tn': tn, 'fp': fp, 'fn': fn, 'tp': tp}
+        conf_mat = metrics.confusion_matrix(all_labels, all_preds)
 
         return avg_loss, metrics_dict, conf_mat
 
     def predict(self, val_loader):
-        """Provide predictions on the validation set.
+        """
+        Provide predictions on the validation set.
 
         Args:
-        ----
             val_loader (Union[torch.utils.data.DataLoader, torch_geometric.loader.DataLoader]): DataLoader for the validation dataset.
-
         Returns:
-        -------
             list: List of predictions.
-
         """
         all_preds = []
         self.model.eval()
@@ -237,26 +260,23 @@ class MulticlassModelTrainer(TorchModelTrainer):
 
                 model_kwargs = self.get_model_kwargs(data)
 
-                outputs = self.model(**model_kwargs)
+                outputs = self.model(**model_kwargs).squeeze(-1)
 
-                probs = F.softmax(outputs, dim=1)
-                _, preds = torch.max(probs, 1)
+                probs = F.sigmoid(outputs)
+                preds = (probs > 0.5).int()
 
                 all_preds.extend(preds.tolist())
 
         return all_preds
 
     def predict_proba(self, val_loader):
-        """Provide the probabilities of the predictions on the validation set.
+        """
+        Provide the probabilities of the predictions on the validation set.
 
         Args:
-        ----
             val_loader (Union[torch.utils.data.DataLoader, torch_geometric.loader.DataLoader]): DataLoader for the validation dataset.
-
         Returns:
-        -------
             list: List of predictions' probabilities.
-
         """
         all_probs = []
         self.model.eval()
@@ -269,10 +289,9 @@ class MulticlassModelTrainer(TorchModelTrainer):
 
                 model_kwargs = self.get_model_kwargs(data)
 
-                outputs = self.model(**model_kwargs)
+                outputs = self.model(**model_kwargs).squeeze(-1)
 
-                probs = F.softmax(outputs, dim=1)
-                # _, preds = torch.max(probs, 1)
+                probs = F.sigmoid(outputs)
 
                 all_probs.extend(probs.tolist())
 
@@ -299,60 +318,42 @@ class MulticlassModelTrainer(TorchModelTrainer):
     def _compute_metrics(self, y_true, y_pred):
         accuracy = metrics.accuracy_score(y_true, y_pred)
         balanced_accuracy = metrics.balanced_accuracy_score(y_true, y_pred)
-        precision_macro = metrics.precision_score(
-            y_true,
-            y_pred,
-            zero_division=0,
-            average="macro",
-            labels=torch.arange(self.model.output_dim),
-        )
-        recall_macro = metrics.recall_score(
-            y_true,
-            y_pred,
-            zero_division=0,
-            average="macro",
-            labels=torch.arange(self.model.output_dim),
-        )
-        f1_macro = metrics.f1_score(
-            y_true,
-            y_pred,
-            zero_division=0,
-            average="macro",
-            labels=torch.arange(self.model.output_dim),
-        )
-        precision_micro = metrics.precision_score(
-            y_true,
-            y_pred,
-            zero_division=0,
-            average="micro",
-            labels=torch.arange(self.model.output_dim),
-        )
-        recall_micro = metrics.recall_score(
-            y_true,
-            y_pred,
-            zero_division=0,
-            average="micro",
-            labels=torch.arange(self.model.output_dim),
-        )
-        f1_micro = metrics.f1_score(
-            y_true,
-            y_pred,
-            zero_division=0,
-            average="micro",
-            labels=torch.arange(self.model.output_dim),
-        )
+        precision = metrics.precision_score(y_true, y_pred, zero_division=0)
+        recall = metrics.recall_score(y_true, y_pred, zero_division=0)
+        f1 = metrics.f1_score(y_true, y_pred, zero_division=0)
         mcc = metrics.matthews_corrcoef(y_true, y_pred)
 
         metrics_dict = {
             "accuracy": accuracy,
             "balanced_accuracy": balanced_accuracy,
-            "precision_macro": precision_macro,
-            "recall_macro": recall_macro,
-            "f1_macro": f1_macro,
-            "precision_micro": precision_micro,
-            "recall_micro": recall_micro,
-            "f1_micro": f1_micro,
+            "precision": precision,
+            "recall": recall,
+            "f1": f1,
             "mcc": mcc,
         }
 
         return metrics_dict
+
+    def pyg_to_onnx(self, featurizer):
+        if self.model.training:
+            self.model.eval()
+            self.model = self.model.cpu()
+
+        dummy_smile = "CCC"
+        dummy_input = featurizer.featurize(dummy_smile)
+        x = dummy_input.x
+        edge_index = dummy_input.edge_index
+        batch = torch.zeros(x.shape[0], dtype=torch.int64)
+        buffer = io.BytesIO()
+        torch.onnx.export(
+            self.model,
+            args=(x, edge_index, batch),
+            f=buffer,
+            input_names=["x", "edge_index", "batch"],
+            dynamic_axes={"x": {0: "nodes"}, "edge_index": {1: "edges"}, "batch": [0]},
+        )
+        onnx_model_bytes = buffer.getvalue()
+        buffer.close()
+        model_scripted_base64 = base64.b64encode(onnx_model_bytes).decode("utf-8")
+
+        return model_scripted_base64
