--- jaqpotpy/models/trainers/binary_trainers/__init__.py
@@ -1,6 +1,0 @@
-"""
-Author: Ioannis Pitoskas (jpitoskas@gmail.com)
-"""
-
-from .binary_model_trainer import BinaryModelTrainer
-from .binary_graph_model_trainer import BinaryGraphModelTrainer
