--- jaqpotpy/models/trainers/multiclass_trainers/__init__.py
@@ -1,6 +1,0 @@
-"""
-Author: Ioannis Pitoskas (jpitoskas@gmail.com)
-"""
-
-from .multiclass_model_trainer import MulticlassModelTrainer
-from .multiclass_graph_model_trainer import MulticlassGraphModelTrainer
