--- jaqpotpy/models/trainers/regression_trainers/__init__.py
@@ -1,6 +1,0 @@
-"""
-Author: Ioannis Pitoskas (jpitoskas@gmail.com)
-"""
-
-from .regression_model_trainer import RegressionModelTrainer
-from .regression_graph_model_trainer import RegressionGraphModelTrainer
